Save analysis in extra function call.
This is much better then when I have to decide at object construction.
I can also decide to save the results after the analysis has run. In the
old way it was not obvious how this should have been done.

This also removes the `_conclude` method since it does nothing now.